(Android) Convert unit-tests to specs using Spek [fixed]
