--- detox/android/detox/src/test/java/com/wix/detox/espresso/EspressoWrapperTest.java
@@ -1,8 +1,0 @@
-package com.wix.detox.espresso;
-
-/**
- * Created by rotemm on 26/12/2016.
- */
-public class EspressoWrapperTest {
-
-}
